+++ src/Kit/DataDisplay/Error/index.tsx
@@ -1,52 +1,65 @@
-import {FunctionComponent} from 'react';
+import {FunctionComponent, ReactElement, Suspense} from 'react';
 import {Empty as AntdEmpty} from 'antd';
 import {IKitError} from './types';
 import {useKitLocale} from '@translation/useKitLocale.ts';
-import {IKitLocale} from '@translation/types';
 import {useKitTheme} from '@theme/useKitTheme';
 import {KitTypography} from '@kit/General';
 import {KitSpace} from '@kit/Layout';
 import cn from 'classnames';
-
-import {error404} from '../../../assets/error/error-404.tsx';
-
 import styles from './styles.module.scss';
+import {KitLoader} from '@kit/Feedback/index.ts';
+import Error401 from '../../../assets/error/error-401.tsx';
+import Error403 from '../../../assets/error/error-403.tsx';
+import Error404 from '../../../assets/error/error-404.tsx';
+import Error500 from '../../../assets/error/error-500.tsx';
+import Error502 from '../../../assets/error/error-502.tsx';
+import Error503 from '../../../assets/error/error-503.tsx';
+import CommonError from '../../../assets/error/common-error.tsx';
 
-const _setLayoutForAntDescription = (
-    {title, description}: Pick<IKitError, 'title' | 'description'>,
-    locale: IKitLocale
-) => {
-    const _description =
-        typeof description === 'string' || typeof description === 'number' ? (
-            <KitTypography.Paragraph>{description}</KitTypography.Paragraph>
-        ) : (
-            description
-        );
-
-    const _errorType = <KitTypography.Paragraph>{locale.Error?.error404}</KitTypography.Paragraph>;
-
-    return (
-        <KitSpace direction="vertical" size="l">
-            <KitTypography.Title className="error-title" level="h1">
-                {title}
-            </KitTypography.Title>
-            {_description}
-            {_errorType}
-        </KitSpace>
-    );
-};
-
-export const KitError: FunctionComponent<IKitError> = ({className, title, description, ...errorProps}) => {
+export const KitError: FunctionComponent<IKitError> = ({
+    className,
+    title,
+    description,
+    httpErrorCode,
+    ...errorProps
+}) => {
     const {appId} = useKitTheme();
     const {locale} = useKitLocale();
 
     const clx = cn(appId, styles['kit-error'], className);
+    const errorImageAlt = `${locale.Error?.error_type}${httpErrorCode}`;
+
+    const _getImageElement = (): ReactElement => {
+        switch (httpErrorCode) {
+            case '401':
+                return <Error401 alt={errorImageAlt} />;
+            case '403':
+                return <Error403 alt={errorImageAlt} />;
+            case '404':
+                return <Error404 alt={errorImageAlt} />;
+            case '500':
+                return <Error500 alt={errorImageAlt} />;
+            case '502':
+                return <Error502 alt={errorImageAlt} />;
+            case '503':
+                return <Error503 alt={errorImageAlt} />;
+            default:
+                return <CommonError />;
+        }
+    };
 
     return (
         <AntdEmpty
             {...errorProps}
-            description={_setLayoutForAntDescription({title, description}, locale)}
-            image={error404}
+            description={
+                <KitSpace direction="vertical" size="l">
+                    <KitTypography.Title className="error-title" level="h1">
+                        {title}
+                    </KitTypography.Title>
+                    {description}
+                </KitSpace>
+            }
+            image={<Suspense fallback={<KitLoader style={{height: '300px'}} />}>{_getImageElement()}</Suspense>}
             className={clx}
         />
     );