+++ src/Kit/DataDisplay/Error/types.d.ts
@@ -1,10 +1,15 @@
 import {EmptyProps} from 'antd';
 import {KitHTMLAttributes} from '../../../types';
-import {ReactNode} from 'react';
+import {ReactElement, ReactNode} from 'react';
 
 type AntdEmptyTypesToOmit = 'description';
+
+type KitHttpErrorCode = '401' | '403' | '404' | '500' | '502' | '503';
+
+type KitErrorImageAlt = ({alt: string}) => ReactElement;
 
 export interface IKitError extends KitHTMLAttributes<HTMLDivElement>, Omit<EmptyProps, AntdEmptyTypesToOmit> {
     title: string;
     description: ReactNode;
+    httpErrorCode?: KitHttpErrorCode;
 }