+++ src/translation/en-US/components/DataDisplay/Error/index.ts
@@ -1,5 +1,5 @@
 import {IKitErrorLocale} from '@translation/types/components/DataDisplay/Error';
 
 export const KitErrorLocale: IKitErrorLocale = {
-    error404: 'Error type: 404'
+    error_type: 'Error type: '
 };